app converted to use boost's bigInt

+++ rsa/rsa/PrimeNumberGenerator.cpp
@@ -1,13 +1,12 @@
 #include "PrimeNumberGenerator.h"
-#include <cmath>
 #include <tuple>
 #include <iostream>
 
 using namespace std;
 
-unsigned long long PrimeNumberGenerator::naive(int numOfBits) {
+cpp_int PrimeNumberGenerator::naive(int numOfBits) {
 	while (true) {
-		unsigned long long p = getRandom(pow(2, numOfBits - 1), pow(2, numOfBits) - 1);
+		cpp_int p = getRandom(pow(cpp_int(2), numOfBits - 1), pow(cpp_int(2), numOfBits) - 1);
 
 		if (p % 2 == 0) {
 			p++;
@@ -18,9 +17,9 @@ unsigned long long PrimeNumberGenerator::naive(int numOfBits) {
 	}
 }
 
-bool PrimeNumberGenerator::naiveTest(unsigned long long p) {
+bool PrimeNumberGenerator::naiveTest(cpp_int p) {
 	while (true) {
-		unsigned long long j = 3;
+		cpp_int j = 3;
 
 		while (j < sqrt(p))
 		{
@@ -33,9 +32,9 @@ bool PrimeNumberGenerator::naiveTest(unsigned long long p) {
 	}
 }
 
-unsigned long long PrimeNumberGenerator::millerRabin(int numOfBits, int s) {
+cpp_int PrimeNumberGenerator::millerRabin(int numOfBits, int s) {
 	while (true) {
-		unsigned long long p = getRandom(pow(2, numOfBits - 1), pow(2, numOfBits) - 1); //not sure
+		cpp_int p = getRandom(pow(cpp_int(2), numOfBits - 1), pow(cpp_int(2), numOfBits) - 1); //not sure
 		if (p % 2 == 0) {
 			p++;
 		}
@@ -46,20 +45,20 @@ unsigned long long PrimeNumberGenerator::millerRabin(int numOfBits, int s) {
 	}
 }
 
-bool PrimeNumberGenerator::millerRabinTest(unsigned long long p, int s) {
+bool PrimeNumberGenerator::millerRabinTest(cpp_int p, int s) {
 	if (p <= 3)
 		return true;
 	if (p % 2 == 0)
 		return false;
 
 	auto tuple = getFactors(p);
-	unsigned long long d = get<0>(tuple);
-	unsigned long long k = get<1>(tuple);
+	cpp_int d = get<0>(tuple);
+	cpp_int k = get<1>(tuple);
 
 	for (int j = 1; j <= s; j++) {
-		unsigned long long a = getRandom(2, p - 2);
+		cpp_int a = getRandom(2, p - 2);
 
-		unsigned long long x = modularExp(a, d, p);
+		cpp_int x = modularExp(a, d, p);
 		if (x == 1) {
 			continue;
 		}
@@ -78,9 +77,9 @@ bool PrimeNumberGenerator::millerRabinTest(unsigned long long p, int s) {
 	return true;
 }
 
-tuple<unsigned long long, unsigned long long> PrimeNumberGenerator::getFactors(unsigned long long p) {
-	unsigned long long  d = p - 1;
-	unsigned long long k = 0;
+tuple<cpp_int, cpp_int> PrimeNumberGenerator::getFactors(cpp_int p) {
+	cpp_int d = p - 1;
+	cpp_int k = 0;
 	while (d % 2 == 0) {
 		d /= 2;
 		k++;
@@ -89,9 +88,9 @@ tuple<unsigned long long, unsigned long long> PrimeNumberGenerator::getFactors(u
 }
 
 
-unsigned long long PrimeNumberGenerator::modularExp(unsigned long long num, unsigned long long pow, unsigned long long mod)
+cpp_int PrimeNumberGenerator::modularExp(cpp_int num, cpp_int pow, cpp_int mod)
 {
-	unsigned long long d = 1;
+	cpp_int d = 1;
 	num = num % mod;
 	while (pow > 0)
 	{
